backend/requirements.txt
alembic
asyncpg
fastapi
psycopg2-binary
pydantic-settings
sqlalchemy
uvicorn
